+++ bigsort/V2/InfixComparer.cs
@@ -8,12 +8,12 @@ namespace BigSort.V2
   /// <summary>
   /// Sorts infixes in the alphabetical order.
   /// </summary>
-  internal class InfixComparer : IComparer<uint>
+  internal class InfixComparer : IComparer<ushort>
   {
     /// <summary>
     /// IComparer.
     /// </summary>
-    public int Compare([AllowNull] uint x, [AllowNull] uint y)
+    public int Compare([AllowNull] ushort x, [AllowNull] ushort y)
     {
       var infix1 = Encoding.Unicode.GetString(BitConverter.GetBytes(x));
       var infix2 = Encoding.Unicode.GetString(BitConverter.GetBytes(y));
